legacy/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
psutil>=5.9.0
colorama>=0.4.6
pybind11>=2.10.3
wmi>=1.5.1
firebase-admin>=6.0.0
